+++ tests/test_ft017th.py
@@ -18,7 +18,6 @@
 import datetime
 import logging
 import queue
-import time
 import unittest.mock
 
 import cc1101
@@ -141,7 +140,8 @@ def test__receive_measurement_unexpected_length():
             sensor._receive_measurement(timeout_seconds=21)
 
 
-def test_receive():
+@pytest.mark.asyncio
+async def test_receive():
     with unittest.mock.patch("cc1101.CC1101"):
         sensor = wireless_sensor.FT017TH(gdo0_gpio_line_name=b"GPIO24")
     measurement_iter = sensor.receive(timeout_seconds=21)
@@ -150,26 +150,45 @@ def test_receive():
     )
     with unittest.mock.patch.object(
         sensor._transceiver, "_wait_for_packet"
-    ) as wait_for_packet_mock, unittest.mock.patch("time.sleep") as sleep_mock:
+    ) as wait_for_packet_mock, unittest.mock.patch(
+        "asyncio.sleep"
+    ) as sleep_mock, unittest.mock.patch.object(
+        sensor, "_parse_transmission", return_value="some-measurement"
+    ) as parse_transmission_mock:
         packet_queue = queue.Queue()
         packet_queue.put(None)
         packet_queue.put(None)
+        packet_queue.put(
+            cc1101._ReceivedPacket(
+                payload=b"too short",
+                rssi_index=0,
+                checksum_valid=True,
+                link_quality_indicator=0,
+            )
+        )
+        packet_queue.put(packet)
         packet_queue.put(packet)
         packet_queue.put(None)
-        packet_queue.put(packet)
         wait_for_packet_mock.side_effect = (
-            lambda gdo0_gpio_line_name, timeout_seconds: packet_queue.get()
+            lambda gdo0_gpio_line_name, timeout_seconds: packet_queue.get_nowait()
         )
-        assert next(measurement_iter) is None  # timeout or error fetching package
-        sleep_mock.assert_called_once_with(1)  # protect against package flood
-        assert next(measurement_iter) is None
-        with unittest.mock.patch.object(
-            sensor, "_parse_transmission", return_value="dummy"
-        ) as parse_transmission_mock:
-            assert next(measurement_iter) == "dummy"
-        assert sleep_mock.call_count == 3
-        sensor._transceiver.__enter__.assert_called_once_with()  # pylint: disable=no-member; false positive
-        assert wait_for_packet_mock.call_count == 3
+        assert sleep_mock.call_count == 0
+        # `anext(…)` introduced in python3.10
+        assert (await measurement_iter.__anext__()) == "some-measurement"
+        assert (  # initial & reconfig after _UnexpectedPacketLengthError
+            # pylint: disable=no-member; false positive
+            sensor._transceiver.__enter__.call_count
+            == 2
+        )
+        assert (  # flood protection before each packet
+            # 1. & 2. timeout or fetching failed
+            # 3. _UnexpectedPacketLengthError
+            # 4. valid
+            # 5. upcoming
+            sleep_mock.call_args_list
+            == [unittest.mock.call(1)] * 5
+        )
+        sleep_mock.reset_mock()
         parse_transmission_mock.assert_called_once()
         (
             parse_transmission_args,
@@ -181,17 +200,30 @@ def test_receive():
             parse_transmission_args[0],
             numpy.array([255, 168] + [0] * 23, dtype=numpy.uint8),
         )
-        assert next(measurement_iter) is None
-        with unittest.mock.patch.object(
-            sensor, "_parse_transmission", return_value="dummy2"
-        ):
-            assert next(measurement_iter) == "dummy2"
-        assert sleep_mock.call_count == 5
+        parse_transmission_mock.reset_mock()
+        assert (await measurement_iter.__anext__()) == "some-measurement"
+        assert (
+            # (5. valid)
+            # 6. upcoming
+            sleep_mock.call_args_list
+            == [unittest.mock.call(1)] * 1
+        )
+        sleep_mock.reset_mock()
+        parse_transmission_mock.assert_called_once()
+        with pytest.raises(queue.Empty):
+            print(await measurement_iter.__anext__())
+        assert (
+            # (6. timeout or fetching failed)
+            # 7. upcoming
+            sleep_mock.call_args_list
+            == [unittest.mock.call(1)] * 2
+        )
     # pylint: disable=no-member; false positive
     sensor._transceiver.unlock_spi_device.assert_not_called()
 
 
-def test_receive_failed_to_decode(caplog):
+@pytest.mark.asyncio
+async def test_receive_failed_to_decode(caplog):
     with unittest.mock.patch("cc1101.CC1101"):
         sensor = wireless_sensor.FT017TH(gdo0_gpio_line_name=b"GPIO24")
     measurement_iter = sensor.receive(timeout_seconds=21)
@@ -213,16 +245,16 @@ def test_receive_failed_to_decode(caplog):
             "dummy measurement",
         ],
     ) as parse_transmission_mock, unittest.mock.patch(
-        "time.sleep"
+        "asyncio.sleep"
     ) as sleep_mock, caplog.at_level(
         logging.DEBUG
     ):
-        assert next(measurement_iter) is None
-        assert next(measurement_iter) is None
-        assert next(measurement_iter) == "dummy measurement"
-    assert wait_for_packet_mock.call_count == 3
-    assert parse_transmission_mock.call_count == 3
-    assert sleep_mock.call_count == 3
+        assert await measurement_iter.__anext__() == "dummy measurement"  # 3rd
+        with pytest.raises(RuntimeError, match=r"StopIteration"):
+            await measurement_iter.__anext__()
+    assert wait_for_packet_mock.call_count == 4
+    assert parse_transmission_mock.call_count == 4
+    assert sleep_mock.call_count == 4
     decode_error_log_records = [
         (r.message, r.exc_info[1])
         for r in caplog.records
@@ -241,7 +273,8 @@ def test_receive_failed_to_decode(caplog):
         )
 
 
-def test_receive_unexpected_packet_length(caplog):
+@pytest.mark.asyncio
+async def test_receive_unexpected_packet_length(caplog):
     with unittest.mock.patch("cc1101.CC1101"):
         sensor = wireless_sensor.FT017TH(gdo0_gpio_line_name=b"GPIO24")
     measurement = wireless_sensor.Measurement(
@@ -257,11 +290,11 @@ def test_receive_unexpected_packet_length(caplog):
             wireless_sensor._UnexpectedPacketLengthError,
             measurement,
         ],
-    ), unittest.mock.patch("time.sleep") as sleep_mock, caplog.at_level(logging.INFO):
+    ), unittest.mock.patch("asyncio.sleep") as sleep_mock, caplog.at_level(
+        logging.INFO
+    ):
         measurement_iter = sensor.receive(timeout_seconds=21)
-        assert next(measurement_iter) is None
-        assert next(measurement_iter) is None
-        assert next(measurement_iter) == measurement
+        assert await measurement_iter.__anext__() == measurement
     assert (
         # pylint: disable=no-member; false positive
         sensor._transceiver.__enter__.call_count
@@ -279,43 +312,11 @@ def test_receive_unexpected_packet_length(caplog):
         ]
         * 2
     )
-    assert sleep_mock.call_count == 3
+    assert sleep_mock.call_count == (2 * 2) + 1
 
 
-@pytest.mark.parametrize("duration_seconds", (1, 21))
-def test__sleep_single(duration_seconds):
-    itr = wireless_sensor.FT017TH._sleep(
-        duration_seconds=duration_seconds, yield_interval=42
-    )
-    with unittest.mock.patch("time.sleep") as sleep_mock:
-        assert next(itr) is None
-        sleep_mock.assert_not_called()
-        assert next(itr) is None
-        sleep_mock.assert_called_once()
-        args, kwargs = sleep_mock.call_args
-        assert len(args) == 1
-        assert args[0] == pytest.approx(duration_seconds, abs=0.1)
-        assert not kwargs
-
-
-def test__sleep_multiple():
-    with unittest.mock.patch("time.sleep") as sleep_mock, unittest.mock.patch(
-        "time.time", side_effect=[100, 100.01, 103, 110, 113, 120, 121, 123]
-    ):
-        for idx, item in enumerate(
-            wireless_sensor.FT017TH._sleep(duration_seconds=23, yield_interval=7)
-        ):
-            assert item is None
-            if idx == 0:
-                sleep_mock.assert_not_called()
-            else:
-                sleep_mock.assert_called_once_with(7)
-                sleep_mock.reset_mock()
-    assert idx + 1 == 3  # pylint: disable=undefined-loop-variable
-    sleep_mock.assert_called_once_with(2)
-
-
-def test_receive_locked(caplog):
+@pytest.mark.asyncio
+async def test_receive_locked(caplog):
     with unittest.mock.patch("cc1101.CC1101"):
         sensor = wireless_sensor.FT017TH(gdo0_gpio_line_name=b"GPIO24")
     measurement = wireless_sensor.Measurement(
@@ -336,35 +337,17 @@ def test_receive_locked(caplog):
     ) as enter_mock, unittest.mock.patch.object(
         sensor, "_receive_measurement", return_value=measurement
     ), unittest.mock.patch(
-        "time.sleep"
+        "asyncio.sleep"
     ) as sleep_mock, caplog.at_level(
         logging.INFO
     ):
-        assert next(measurement_iter) is None  # 1st BlockingIOError
-        sleep_mock.assert_not_called()
-        assert next(measurement_iter) is None  # 1st BlockingIOError
-        sleep_mock.assert_called_once()
-        assert sleep_mock.call_args[0][0] == pytest.approx(2, abs=0.01)
-        sleep_mock.reset_mock()
-        with unittest.mock.patch("time.time", return_value=time.time() + 3):
-            assert next(measurement_iter) is None  # 2nd BlockingIOError
-        sleep_mock.assert_called_once_with(0)  # remaining lock wait of 1st error
-        sleep_mock.reset_mock()
-        with unittest.mock.patch("time.time", return_value=time.time() + 8):
-            assert next(measurement_iter) is None  # 3rd BlockingIOError
-        sleep_mock.assert_called_once_with(0)  # remaining lock wait of 2nd error
-        sleep_mock.reset_mock()
-        with unittest.mock.patch("time.time", return_value=time.time() + 17):
-            assert next(measurement_iter) is measurement
-        assert sleep_mock.call_count == 2
-        assert sleep_mock.call_args_list[0][0][0] == 0  # remaining lock wait
-        assert sleep_mock.call_args_list[1][0][0] == 1  # flood protection
-        sleep_mock.reset_mock()
-        with unittest.mock.patch.object(
-            sensor, "_receive_measurement", side_effect=BlockingIOError
-        ):
-            assert next(measurement_iter) is None
-        sleep_mock.assert_called_once_with(1)  # flood protection
+        assert await measurement_iter.__anext__() == measurement
+        assert sleep_mock.call_args_list == [
+            unittest.mock.call(2),
+            unittest.mock.call(4),
+            unittest.mock.call(8),
+            unittest.mock.call(1),  # flood protection
+        ]
     assert enter_mock.call_count == 4
     assert caplog.record_tuples == [
         (
@@ -372,11 +355,12 @@ def test_receive_locked(caplog):
             logging.INFO,
             f"SPI device locked, waiting {seconds} seconds",
         )
-        for seconds in [2, 4, 8, 2]
+        for seconds in [2, 4, 8]
     ]
 
 
-def test_receive_no_reconfiguring(caplog):
+@pytest.mark.asyncio
+async def test_receive_no_reconfiguring(caplog):
     with unittest.mock.patch("cc1101.CC1101"):
         sensor = wireless_sensor.FT017TH(gdo0_gpio_line_name=b"GPIO24")
     measurement = wireless_sensor.Measurement(
@@ -386,10 +370,10 @@ def test_receive_no_reconfiguring(caplog):
     )
     with unittest.mock.patch.object(
         sensor, "_receive_measurement", return_value=measurement
-    ), unittest.mock.patch("time.sleep"), caplog.at_level(logging.INFO):
+    ), unittest.mock.patch("asyncio.sleep"), caplog.at_level(logging.INFO):
         measurement_iter = sensor.receive(timeout_seconds=21)
         for _ in range(3):
-            assert next(measurement_iter) == measurement
+            assert await measurement_iter.__anext__() == measurement
     assert (
         # pylint: disable=no-member; false positive
         sensor._transceiver.__enter__.call_count
@@ -398,7 +382,8 @@ def test_receive_no_reconfiguring(caplog):
     assert not caplog.record_tuples
 
 
-def test_receive_unlock_spi_device(caplog):
+@pytest.mark.asyncio
+async def test_receive_unlock_spi_device(caplog):
     with unittest.mock.patch("cc1101.CC1101") as transceiver_class_mock:
         sensor = wireless_sensor.FT017TH(
             gdo0_gpio_line_name=b"GPIO24", unlock_spi_device=True
@@ -411,14 +396,51 @@ def test_receive_unlock_spi_device(caplog):
     )
     with unittest.mock.patch.object(
         sensor, "_receive_measurement", return_value=measurement
-    ), unittest.mock.patch("time.sleep"), caplog.at_level(logging.DEBUG):
+    ), unittest.mock.patch("asyncio.sleep"), caplog.at_level(logging.DEBUG):
         measurement_iter = sensor.receive(timeout_seconds=21)
         for _ in range(3):
-            assert next(measurement_iter) == measurement
+            assert await measurement_iter.__anext__() == measurement
     transceiver_class_mock().unlock_spi_device.assert_called_once_with()
     assert len(caplog.records) == 2
     assert caplog.record_tuples[1] == (
         "wireless_sensor",
         logging.DEBUG,
         "unlocked SPI device",
+    )
+
+
+@pytest.mark.asyncio
+async def test_receive_timeout(caplog):
+    with unittest.mock.patch("cc1101.CC1101"):
+        sensor = wireless_sensor.FT017TH(
+            gdo0_gpio_line_name=b"GPIO24", unlock_spi_device=False
+        )
+    measurement = wireless_sensor.Measurement(
+        decoding_timestamp=datetime.datetime.now(),
+        temperature_degrees_celsius=21.0,
+        relative_humidity=42.0,
+    )
+    with unittest.mock.patch.object(
+        sensor,
+        "_receive_measurement",
+        side_effect=[
+            measurement,
+            measurement,
+            None,
+            measurement,
+            None,
+            None,  # timeout
+            measurement,
+        ],
+    ), caplog.at_level(logging.WARNING):
+        measurement_iter = sensor.receive(timeout_seconds=2)
+        for _ in range(3):
+            assert await measurement_iter.__anext__() == measurement
+        with pytest.raises(StopAsyncIteration):
+            await measurement_iter.__anext__()
+    assert len(caplog.records) == 1
+    assert caplog.record_tuples[0] == (
+        "wireless_sensor",
+        logging.WARNING,
+        "timeout waiting for packet",
     )